Backport rs6000: Adjust -fpatchable-function-entry* support for dual entry
Original message:
As the discussion in PR112980, although the current
implementation for -fpatchable-function-entry* conforms
with the documentation (making N NOPs be consecutive),
it's inefficient for both kernel and userspace livepatching
(see comments in PR for the details).

So this patch is to change the current implementation by
emitting the "before" NOPs before global entry point and
the "after" NOPs after local entry point.  The new behavior
would not keep NOPs to be consecutive, so the documentation
is updated to emphasize this

Backport to gcc-13

Authored-by: Kewen Lin <linkw@linux.ibm.com>
Backported-by: Giuliano Belinassi <gbelinassi@suse.de>

	PR target/112980

gcc/ChangeLog:

	* config/rs6000/rs6000-logue.cc (rs6000_output_function_prologue):
	Adjust the handling on patch area emitting with dual entry, remove
	the restriction on "before" NOPs count, not emit "before" NOPs any
	more but only emit "after" NOPs.
	* config/rs6000/rs6000.cc (rs6000_print_patchable_function_entry):
	Adjust by respecting cfun->machine->stop_patch_area_print.
	(rs6000_elf_declare_function_name): For ELFv2 with dual entry, set
	cfun->machine->stop_patch_area_print as true.
	* config/rs6000/rs6000.h (struct machine_function): Remove member
	global_entry_emitted, add new member stop_patch_area_print.
	* doc/invoke.texi (option -fpatchable-function-entry): Adjust the
	documentation for PowerPC ELFv2 dual entry.

gcc/testsuite/ChangeLog:

	* c-c++-common/patchable_function_entry-default.c: Adjust.
	* gcc.target/powerpc/pr99888-4.c: Likewise.
	* gcc.target/powerpc/pr99888-5.c: Likewise.
	* gcc.target/powerpc/pr99888-6.c: Likewise.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>