Drop references to externalized functions on initializers
Currently, the externalization method involves in redeclaring a
function `func1` as `*klpe_func1`, which is then filled with the
address of the original function when libpulp is loaded.  Unfortunately
there is no way we can know this at compile time so we have to remove
the reference in this case.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>